requirements.txt
Django==1.6.11
Pillow==2.5.3
South==0.8.4
akismet==0.2.0
argparse==1.2.1
beautifulsoup4==4.3.2
django-auth-ldap==1.2.0
django-blog-zinnia==0.14.1
-e git+https://github.com/dwaiter/django-ckeditor.git@391ad930ff8f51be7f91943b68ee8df8c58f18ae#egg=django_ckeditor-master
django-localeurl==2.0.1
django-model-utils==2.0.3
django-mptt==0.6.1
django-oembed==0.1.3
django_sendmail_backend==0.1.2
django-tagging==0.3.4
django-xmlrpc==0.1.5
gevent==1.0.1
greenlet==0.4.2
gunicorn==18.0
git+ssh://git@gc-web.de:222/gcweb/imagetools@0d75f39#egg=imagetools
git+ssh://git@gc-web.de:222/gcweb/simplecms@74ce596#egg=simplecms
wsgiref==0.1.2
ipython==1.1.0
psycopg2==2.5.3
pyparsing==2.0.1
python-ldap==2.4.15
pytz==2014.4
setproctitle==1.1.8
django-scheduler==0.7.4
